+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Update Forge2D docs
  - Update PR template with removal of develop branch
  - Translate README to Russian
+ - Split up Component and PositionComponent to BaseComponent
 
 ## 1.0.0-rc2
  - Improve IsometricTileMap and Spritesheet classes